% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-data.R
\docType{data}
\name{ants}
\alias{ants}
\title{Ant Biodiversity Data}
\format{A data.frame with 82 rows (species) and 5 columns (covertypes):
\describe{
  \item{species}{Ant species (family)}
  \item{Primary.Forest}{Primary forest type}
  \item{Abandoned.cacao.plantations}{Abandoned cacao plantations type}
  \item{Productive.cacao.plantations}{Active cacao plantations type}
  \item{Banana.plantations}{Active banana plantations type}
}}
\source{
\url{http://www.tiem.utk.edu/~gross/bioed/bealsmodules/shannonDI.html}
}
\description{
Roth et al., (1994) Costa Rican ant diversity data
}
\references{
Roth, D. S., I. Perfecto, and B. Rathcke (1994) The effects of management systems on 
            ground-foraging ant diversity in Costa Rica. Ecological Applications 4(3):423-436.
}
