% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgls_urls.R
\name{cgls_urls}
\alias{cgls_urls}
\title{Provide URL's for Copernicus Global Land Service datasets}
\usage{
cgls_urls(...)
}
\arguments{
\item{...}{not used}
}
\description{
Returns URL's of a product/version/resolution
}
\details{
Given the changes to ESA-Copenricus data distribution, using digests is
no longer reliable. Plese use the Sentinel Hub or a STAC API for data 
acccess
}
