% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.data.R
\docType{data}
\name{cor.data}
\alias{cor.data}
\title{Various correlation structures}
\format{
A list object with 4 elements containing data.frames:
\describe{
  \item{example 1}{two columns with nonlinear wave function relationship}
  \item{example 2}{two columns with simple nonlinear relationship}
  \item{example 3}{two columns with nonlinear multi-level wave function relationship}
  \item{example 4}{4 columns with first two having linear relationship}
}
}
\description{
linear and nonlinear correlated data examples

a list object with various linear and nonlinear correlation structures
}
