% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet.tiles.R
\name{daymet.tiles}
\alias{daymet.tiles}
\title{DAYMET Tile ID's}
\usage{
daymet.tiles(...)
}
\arguments{
\item{...}{ignored}
}
\value{
Vector of DAYMET tile IDS or if sp = TRUE a sp class SpatialPolygonsDataFrame
}
\description{
Returns a vector of DAYMET tile id's within a specified extent
}
\note{
Function accepts sp, raster or extent class object or bounding coordinates. 
All input must be in the same projection as the tile index SpatialPolygonsDataFrame. 
The library includes the DAYMAT tile index "DAYMET_tiles" which can be add using 
data(), see examples.
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
