% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAYMET_tiles-data.R
\docType{data}
\name{DAYMET_tiles}
\alias{DAYMET_tiles}
\title{DAYMET climate tile index}
\format{
An sp SpatialPolygonsDataFrame with 404 features (rows) and 6 columns (columns):
\describe{
  \item{Id}{Tile Index Identification}
  \item{Area}{Area of each tile}
  \item{XMin}{Minimum x geographic decimal degree coordinate}
  \item{XMax}{Maximum x geographic decimal degree coordinate}
  \item{YMin}{Minimum y geographic decimal degree coordinate}
  \item{yMax}{Maximum y geographic decimal degree coordinate}
}
}
\source{
\url{https://daymet.ornl.gov/}
}
\description{
Polygon tile index for DAYMET climate data
}
