% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence.R
\name{divergence}
\alias{divergence}
\title{divergence}
\usage{
divergence(x, y, type = c("Kullback-Leibler", "cross-entropy"))
}
\arguments{
\item{x}{a vector of integer values, defining observed}

\item{y}{a vector of integer values, defining estimates}

\item{type}{Type of divergence statistic c("Kullback-Leibler", 
"cross-entropy")}
}
\value{
single value vector with divergence statistic
}
\description{
Kullback-Leibler Divergence (Cross-entropy)
}
\examples{
x <- round(runif(10,1,4),0)
y <- round(runif(10,1,4),0)

divergence(x, y) 
divergence(x, y, type = "cross-entropy") 

}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
