% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.daymet.R
\name{download.daymet}
\alias{download.daymet}
\title{Download DAYMET}
\usage{
download.daymet(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Batch download of daily gridded DAYMET climate data
}
\details{
DAYMET website: \url{http://daymet.ornl.gov},
path structure: /year/tile_year/file.nc
}
\references{
Thornton P.E., S.W. Running and M.A. White (1997) Generating surfaces of daily 
  meteorological variables over large regions of complex terrain. Journal of
  Hydrology 190: 214-251.

Thornton, P.E. and S.W. Running (1999) An improved algorithm for estimating 
  incident daily solar radiation from measurements of temperature, humidity, 
  and precipitation. Agriculture and Forest Meteorology. 93:211-228.

Thornton, P.E., H. Hasenauer and M.A. White (2000) Simultaneous estimation 
  of daily solar radiation and humidity from observed temperature and 
  precipitation: An application over complex terrain in Austria. 
  Agricultural and Forest Meteorology 104:255-271.
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
