% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev-data.R
\docType{data}
\name{elev}
\alias{elev}
\title{Elevation raster}
\format{
A raster RasterLayer class object:
\describe{
  \item{resoultion}{5 arc-minute 0.00833 (10000m)}
  \item{nrow}{264}
  \item{ncol}{564}
  \item{ncell}{148896}
  \item{xmin}{5.9}
  \item{xmax}{10.6}
  \item{ymin}{45.7}
  \item{ymax}{47.9}
  \item{proj4string}{+proj=longlat +ellps=WGS84}
}
}
\source{
\url{http://www.diva-gis.org/Data}
}
\description{
elevation raster of Switzerland
}
