% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.lmetrics.R
\name{focal.lmetrics}
\alias{focal.lmetrics}
\title{Focal landscape metrics}
\usage{
focal.lmetrics(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version 
of the function}
}
\description{
Calculates a variety of landscape metrics on  
             integer rasters using focal approach
}
\examples{
\dontrun{
library(landscapemetrics)
library(raster)

data(landscape)

s <- matrix(1, nrow = 3, ncol = 3)
( result <- do.call(stack, window_lsm(landscape, window = s, 
                  what = c("lsm_l_pr", "lsm_l_joinent"))) )
  plot(result)
}
}
