% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzySum.R
\name{fuzzySum}
\alias{fuzzySum}
\title{Fuzzy Sum}
\usage{
fuzzySum(x)
}
\arguments{
\item{x}{Vector of values to apply fuzzy sum}
}
\value{
Value of fuzzy sum
}
\description{
Calculates the fuzzy sum of a vector
}
\note{
The fuzzy sum is an increasing linear combination of values. 
This can be used to sum probabilities or results of multiple 
density functions.
}
\examples{
p = c(0.8,0.76,0.87)
  fuzzySum(p)
  sum(p)

p = c(0.3,0.2,0.1)
  fuzzySum(p)
  sum(p)  

}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
