% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.info.R
\name{grid.info}
\alias{grid.info}
\title{Grid Information}
\usage{
grid.info(lats, cellsize, r = 6378137)
}
\arguments{
\item{lats}{geographic coords}

\item{cellsize}{cellsize}

\item{r}{radius of earth}
}
\description{
Grid Information from Geographic (lat lon) Projections
}
\note{
Since spatial grids in geographic projections do not have equal area or
perimeters, grid.info extracts perimeter & area related information
for latitudinal bands with differing longitudinal widths. Outputs
lengths are in m using Vincenty's equation distance and areas in m2.
Surface areas are calculated summing surface areas of spherical polygons as
estimated using l'Huiller's formula.
}
