% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hli.R
\name{hli}
\alias{hli}
\title{Heat Load Index}
\usage{
hli(x)
}
\arguments{
\item{x}{raster object}
}
\value{
raster class object of McCune & Keon (2002) Heat Load Index
}
\description{
Calculates the McCune & Keon (2002) Heat Load Index
}
\note{
Describes A southwest facing slope should have warmer temperatures than a southeast facing slope, even though the amount of solar radiation they receive is equivalent. The McCune and Keon (2002) method accounts for this by "folding" the aspect so that the highest values are southwest and the lowest values are northeast. Additionally, this method account for steepness of slope, which is not addressed in most other aspect rescaling equations. HLI values range from 0 (coolest) to 1 (hottest).
}
\examples{
  library(raster)
  data(elev)
  heat.load <- hli(elev)
    plot(heat.load, main="Heat Load Index") 
    
}
\references{
McCune, B., and D. Keon (2002) Equations for potential annual direct incident radiation and heat load index. Journal of Vegetation Science. 13:603-606.
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
