% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.whole.R
\name{is.whole}
\alias{is.whole}
\title{is.whole}
\usage{
is.whole(a, tol = 0.0000001)
}
\arguments{
\item{a}{A numeric vector to evaluate, only first element will be evaluated}

\item{tol}{numeric ≥ 0, differences smaller than tolerance are not reported}
}
\value{
A Boolean indicating if number is whole or float
}
\description{
Boolean for evaluating whole numbers
}
\examples{
is.whole( 1 )
is.whole( 1.5 )
is.whole( 0.5 )

}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
