% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde2D.R
\name{kde.2D}
\alias{kde.2D}
\title{2-dimensional kernel density estimate}
\usage{
kde.2D(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
Calculates 2-dimensional kernel density estimate over specified extent
}
