% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries}
\alias{libraries}
\title{libraries}
\usage{
libraries(
  x,
  add = FALSE,
  install = TRUE,
  check.source = TRUE,
  repository = NULL,
  lib = NULL
)
}
\arguments{
\item{x}{Quoted package name(s)}

\item{add}{(FALSE/TRUE) Add package(s) to current R environment using require}

\item{install}{(TRUE/FALSE) If missing from library, install package}

\item{check.source}{(TRUE/FALSE) Check for recent source compile}

\item{repository}{Use specific repository.}

\item{lib}{A local R library path, default is the .Library global}
}
\value{
installed package(s) optionally added to environment
}
\description{
Checks, installs and optionally adds libraries to current environment
}
\note{
This function will check if a package is installed in the R library and 
optionally install the package and add it to the active environment using require.
A check on the package availability in the defined mirror is also performed. This 
can be used to manage package installs but, also to merely check the install and add
}
\examples{
\dontrun{
# Install packages, add to namespace environment and suppress source check
libraries( c("sp","spdep"), add=TRUE, check.source=FALSE)

# Check package(s) install and add to namespace environment but, 
#   do not install if missing
libraries( c("sp","spdep","blah"), install=FALSE, add=TRUE)
} 

}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
