% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.values.R
\name{nth.values}
\alias{nth.values}
\title{Nth values}
\usage{
nth.values(x, N = 2, smallest = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{N}{Number of (Nth) values returned}

\item{smallest}{(FALSE/TRUE) Return the highest, else smallest values}
}
\value{
Numeric vector of Nth values
}
\description{
Returns the Nth highest or lowest values in a vector
}
\note{
This function returns n lowest or highest elements in a vector
}
\examples{
nth.values(1:20, N=3, smallest = TRUE)                 
nth.values(1:20, N=3)

}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
