% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oli.aws.R
\name{oli.aws}
\alias{oli.aws}
\title{Query AWS-OLI}
\usage{
oli.aws(...)
}
\arguments{
\item{...}{not used}
}
\description{
Query of Amazon AWS OLI-Landsat 8 cloud service
}
\details{
Given the changes to AWS Registry of Open Data, using the data  
digest is no longer reliable. Plese use the AWS Data Exchange 
or a STAC API for data acccess
}
