% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.effect.size.R
\name{plot.effect.size}
\alias{plot.effect.size}
\title{Plot effect size}
\usage{
\method{plot}{effect.size}(x, ...)
}
\arguments{
\item{x}{A effect.size object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot function for effect.size object
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
