% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_extract.R
\name{polygon_extract}
\alias{polygon_extract}
\title{polygon raster extract}
\usage{
polygon_extract(...)
}
\arguments{
\item{...}{arguments passed to terra::extract}
}
\value{
NA
}
\description{
Fast method for extracting raster values to polygons
}
\examples{
 \dontrun{
  terra::extract()
}

}
