% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyPerimeter.R
\name{polyPerimeter}
\alias{polyPerimeter}
\title{Polygon perimeter}
\usage{
polyPerimeter(x)
}
\arguments{
\item{x}{sf POLYGON class object}
}
\value{
A vector of polygon perimeters in projection units
}
\description{
Calculates the perimeter length(s) for a polygon object
}
\examples{
library(sf)
  polys <- st_read(system.file("shape/nc.shp", package="sf"))
    polys <- suppressWarnings(st_cast(polys[c(10,100),], "POLYGON"))
 
 polyPerimeter(polys)

}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
