% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.effect.size.R
\name{print.effect.size}
\alias{print.effect.size}
\title{Print effect size}
\usage{
\method{print}{effect.size}(x, ...)
}
\arguments{
\item{x}{Object of class effect.size}

\item{...}{Ignored}
}
\value{
Prints the output data.frame contaning; effect size with upper and lower confidence 
and, mean and sd by group
}
\description{
print method for class "effect.size"
}
