% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.poly.trend.R
\name{print.poly.trend}
\alias{print.poly.trend}
\title{Print poly_trend}
\usage{
\method{print}{poly.trend}(x, ...)
}
\arguments{
\item{x}{Object of class poly.trend}

\item{...}{Ignored}
}
\value{
Prints trend model summary, order and trend confidence intervals
}
\description{
print method for class "poly.trend"
}
