% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pu-data.R
\docType{data}
\name{pu}
\alias{pu}
\title{Biodiversity Planning Units}
\format{
A sp SpatialPolygonsDataFrame with 5919 rows and 46 variables:
\describe{
  \item{UNIT_ID}{Unique planning unit ID}
  \item{DR_Dr_A}{Biodiversity target}
  \item{DR_Dr_L}{Biodiversity target}
  \item{Ht_Dr_A}{Biodiversity target}
  \item{Ht_Dr_L}{Biodiversity target}
  \item{DR_Ms_A}{Biodiversity target}
  \item{DR_Ms_L}{Biodiversity target}
  \item{Ht_Ms_L}{Biodiversity target}
  \item{DR_LM_M}{Biodiversity target}
  \item{H_LM_M_L}{Biodiversity target}
  \item{H_LM_R_L}{Biodiversity target}
  \item{DR_LM_R_L}{Biodiversity target}
  \item{DR_Rn_L}{Biodiversity target}
  \item{DR_LM_R_S}{Biodiversity target}
  \item{DR_Rn_S}{Biodiversity target}
  \item{DR_Ms_S}{Biodiversity target}
  \item{Ht_Ms_A}{Biodiversity target}
  \item{DR_Ms_E}{Biodiversity target}
  \item{DR_Ms_I}{Biodiversity target}
  \item{DR_Rn_E}{Biodiversity target}
  \item{DR_Rn_I}{Biodiversity target}
  \item{H_LM_R_E}{Biodiversity target}
  \item{Ht_Ms_E}{Biodiversity target}
  \item{Ht_Rn_E}{Biodiversity target}
  \item{DR_Rn_A}{Biodiversity target}
  \item{Ht_Rn_A}{Biodiversity target}
  \item{Ht_Rn_I}{Biodiversity target}
  \item{Ht_Dr_E}{Biodiversity target}
  \item{Ht_Ms_S}{Biodiversity target}
  \item{Ht_Dr_S}{Biodiversity target}
  \item{Ht_Rn_L}{Biodiversity target}
  \item{Ht_Th_A}{Biodiversity target}
  \item{Ht_Th_L}{Biodiversity target}
  \item{Ht_Th_S}{Biodiversity target}
  \item{Ht_Dr_U}{Biodiversity target}
  \item{Ht_Dr_I}{Biodiversity target}
  \item{Ht_Ms_I}{Biodiversity target}
  \item{H_LM_M_A}{Biodiversity target}
  \item{H_LM_M_E}{Biodiversity target}
  \item{H_LM_R_A}{Biodiversity target}
  \item{H_LM_M_S}{Biodiversity target}
  \item{H_LM_R_I}{Biodiversity target}
  \item{H_LM_R_S}{Biodiversity target}
  \item{Ht_Rn_S}{Biodiversity target}
  \item{Ht_Ms_U}{Biodiversity target}
  \item{Ht_Rn_U}{Biodiversity target}
}
}
\source{
\url{https://www.conservationgateway.org}
}
\description{
Subset of biodiversity planning units for Haiti ecoregional spatial reserve plan
}
\references{
Evans, J.S., S.R. Schill, G.T. Raber (2015) A Systematic Framework for Spatial Conservation Planning and Ecological Priority Design in St. Lucia, Eastern Caribbean. Chapter 26 in Central American Biodiversity : Conservation, Ecology and a Sustainable Future. F. Huettman (eds). Springer, NY.
}
