% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.ext.R
\name{rm.ext}
\alias{rm.ext}
\title{Remove extension}
\usage{
rm.ext(x)
}
\arguments{
\item{x}{A character vector representing a file with extension}
}
\value{
The file name with extension and file path stripped off
}
\description{
Removes file extension (and path) from string
}
\examples{
rm.ext("C:/path/file.txt")

}
