% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.line.R
\name{sample.line}
\alias{sample.line}
\title{Sample Lines}
\usage{
sample.line(...)
}
\arguments{
\item{...}{arguments passed to sf::st_sample}
}
\value{
NA
}
\description{
Creates a sample for each line in a 
             sf LINESTRING class object
}
\examples{
 \dontrun{
  sf::sf_sample()
}

}
