% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.poly.R
\name{sample.poly}
\alias{sample.poly}
\title{Sample Polygons}
\usage{
sample.poly(...)
}
\arguments{
\item{...}{arguments passed to sf::st_sample}
}
\value{
NA
}
\description{
Creates an equal sample of n for each polygon in an 
             sp Polygon class object
}
\examples{
 \dontrun{
  sf::sf_sample()
}

}
