% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_dissolve.R
\name{sf_dissolve}
\alias{sf_dissolve}
\title{Dissolve polygons}
\usage{
sf_dissolve(x, y = NULL)
}
\arguments{
\item{x}{An sf POLYGON or MULTIPOLYGON object}

\item{y}{An attribute in x to dissolve by, default is NULL}
}
\value{
A dissolved POLYGON or MULTIPOLYGON object
}
\description{
Dissolve polygon feature calss
}
\note{
If a dissolve attribute is defined, the result will be a 
MULTIPOLYGON with the grouping attribute column. If y=NULL all polygons
will be dissolved into a single attribute, unless there is spatial
discontinuity (eg., gaps) in the data.
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>

library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
  nc$group <- ifelse(nc$CNTY_ <= 1902, 1, 
                ifelse(nc$CNTY_ > 1902 & nc$CNTY_ <= 1982, 2, 
				     ifelse(nc$CNTY_ > 1982, 3, NA))) 

# Dissolve by group attribute
d <- sf_dissolve(nc, "group")
  plot(st_geometry(nc), border="grey")
    plot(st_geometry(d), border="red", col=NA, 
         lwd=2, add=TRUE) 

# Dissolve all polygons
d <- sf_dissolve(nc)
  plot(st_geometry(nc), border="grey")
    plot(st_geometry(d), border="red", col=NA, 
         lwd=2, add=TRUE)
}
