% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEcoNews.R
\name{spatialEcoNews}
\alias{spatialEcoNews}
\title{spatialEco news}
\usage{
spatialEcoNews(...)
}
\arguments{
\item{...}{not used}
}
\value{
Shows package NEWS file
}
\description{
Displays release notes
}
