% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.effect.size.R
\name{summary.effect.size}
\alias{summary.effect.size}
\title{Summarizing effect size}
\usage{
\method{summary}{effect.size}(object, ...)
}
\arguments{
\item{object}{Object of class effect.size}

\item{...}{Ignored}
}
\value{
Prints the output data.frame contaning; effect size with upper and lower confidence 
and, mean and sd by group
}
\description{
Summary method for class "effect.size".
}
