% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.loess.boot.R
\name{summary.loess.boot}
\alias{summary.loess.boot}
\title{Summarizing Loess bootstrap models}
\usage{
\method{summary}{loess.boot}(object, ...)
}
\arguments{
\item{object}{Object of class loess.boot}

\item{...}{Ignored}
}
\description{
Summary method for class "loess.boot".
}
