% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.comparison.R
\name{class.comparison}
\alias{class.comparison}
\title{Class comparison between two nominal rasters}
\usage{
class.comparison(...)
}
\arguments{
\item{...}{arguments passed to raster.change}
}
\value{
NA
}
\description{
Compares two categorical rasters using Cohen's Kappa (d) 
             or paired t-test statistic(s)
}
\examples{
 \dontrun{
  raster.change()
}

}
