% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexHull.R
\name{convexHull}
\alias{convexHull}
\title{Alpha Convex Hull}
\usage{
convexHull(x, alpha = 250000)
}
\arguments{
\item{x}{SpatialPoints or SpatialPointsDataFrame object}

\item{alpha}{Alpha parameter for adjusting boundary tension}
}
\value{
SpatialPolygons object
}
\description{
Calculates a convex hull on a point feature class using the Pateiro-Lopez (2009) Alphahull model
}
\note{
This method provides flexibility over traditional convex functions in that the the alpha parameter 
      can be adjusted to relax or increase tension between boundary-edge points
      Due to licensing constraints associated with the alphahull package, this function is not available 
      in the CRAN release. The function must be called from the package NAMESPACE using: 
      spatialEco:::convexHull
}
\examples{
 library(sp)
 data(meuse)
  coordinates(meuse) = ~x+y
 a <- convexHull(meuse, alpha=100000)
   plot(a)
     points(meuse, pch=19)
 
 # Convert SpatialLinesDataFrame to SpatialPolygonsDataFrame
 library(sf)
 a <- sf::st_as_sf(a) 
 a <- sf::st_polygonize(a)
 class( a <- as(a, "Spatial") )
   plot(a)
 
 # Test multiple alpha values
  par(mfcol=c(2,2))
    for (a in c(500, 1500, 5000, 100000)) {
    ch <- convexHull(meuse, alpha = a)
      plot(ch)
       points(meuse, pch=19)
         title( paste0("alpha=", a))		 
    }

}
\references{
Pateiro-Lopez & Rodriguez-Casal (2009) Generalizing the Convex Hull of a Sample: The R Package alphahull. Journal of Statistical Software 34(5):1-28 http://www.jstatsoft.org/v34/i05/paper
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
