% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.hansen.R
\name{download.hansen}
\alias{download.hansen}
\title{Download Hansen Forest 2000-2013 Change}
\usage{
download.hansen(...)
}
\arguments{
\item{...}{Nonexistent parameters}
}
\description{
Download of Hansen Global Forest Change 2000-2013
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
