% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hli.R
\name{hli}
\alias{hli}
\title{Heat Load Index}
\usage{
hli(
  aspect,
  slope,
  latitude = NULL,
  direct = FALSE,
  scaled = FALSE,
  units = c("degrees", "radians"),
  hemisphere = c("northern", "southern"),
  force.hemisphere = TRUE,
  equation = c(1, 2, 3)
)
}
\arguments{
\item{aspect}{terra SpatRaster class object of aspect}

\item{slope}{terra SpatRaster class object of slope}

\item{latitude}{Vector of center latitude of raster, if NULL automatical calculated}

\item{direct}{(FALSE/TRUE) default is "heatload", alternative is "radiation"}

\item{scaled}{(FALSE/TRUE) default is "heatload", alternative is "radiation"}

\item{units}{Default is in "degrees", alternative is "radians"}

\item{hemisphere}{Define if in "southern" or "northern" hemisphere, default is northern}

\item{force.hemisphere}{Force hemisphere to southern or northern}

\item{equation}{Which equation to use (1,2,3), default is 1 covering all slopes <= 90
and latitudes 0-60N}
}
\value{
terra SpatRaster class object of McCune & Keon (2002) Heat Load Index
}
\description{
Calculates the McCune & Keon (2002) Heat Load Index
}
\details{
Describes A southwest facing slope should have warmer temperatures than a 
southeast facing slope, even though the amount of solar radiation they receive 
is equivalent. The McCune and Keon (2002) method accounts for this by "folding" 
the aspect so that the highest values are southwest and the lowest values are  
northeast. Additionally, this method accounts for steepness of slope, which is 
not addressed in most other aspect rescaling equations. 

The equations follow McCune (2007) and support northern and southern hemisphere 
calculations. The . If a country is split at the McCune (2007) folded aspect corrections 
for northern and southern latitudes are included. If you are bounding the equator you can 
use the force.hemisphere argument to choose which equation to use. Valid values for this 
argument are "southern" and "northern" with the default "none". Metic values are returned 
in ln(Rad, MJ  cm–2 yr–1) however, if scale is TRUE they are returned in an arithmetic scale.     

Equations - the equation number (1, 2 or 3); default is 1 (the most general one). The three 
equations have slightly different uses. Eq. 1 (default in the function below) is broadest in 
the application, covering all slopes <= 90 degrees in steepness at latitudes 0-60N, but has the 
lowest precision. Eq. 2 increases the precision by excluding slopes steeper than 60, an 
inconsequential omission for almost all data sets. Eq. 3 uses only three parameters to 
produce a slightly stronger model but is applicable only to latitudes 30-60N (ie., not for the 
tropical and subtropical region).
}
\examples{
  library(terra)
  elev <- rast(system.file("extdata/elev.tif", package="spatialEco"))
  sa <- terrain(elev, v = c("slope", "aspect")) 
  heat.load <- hli(sa[[2]], sa[[1]])
    plot(heat.load, main="Heat Load Index", smooth=TRUE) 
   
  heat.load2 <- hli(sa[[2]], sa[[1]], equation = 2)
    plot(heat.load2, main="Heat Load Index eq 2", smooth=TRUE) 

}
\references{
McCune, B., and D. Keon (2002) Equations for potential annual direct 
  incident radiation and heat load index. Journal of Vegetation 
  Science. 13:603-606.

McCune, B. (2007). Improved estimates of incident radiation and heat load 
  using non-parametric regression against topographic variables. Journal 
  of Vegetation Science 18:751-754.
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
