% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp.R
\name{hsp}
\alias{hsp}
\title{Hierarchical Slope Position}
\usage{
hsp(x, min.scale = 3, max.scale = 27, inc = 4, win = "rectangle",
  normalize = FALSE)
}
\arguments{
\item{x}{Object of class raster (requires integer raster)}

\item{min.scale}{Minimum scale (window size)}

\item{max.scale}{Maximum scale (window size)}

\item{inc}{Increment to increase scales}

\item{win}{Window type, options are "rectangle" or "circle"}

\item{normalize}{Normalize results to 0-1 scale (FALSE | TRUE)}
}
\value{
raster class object
}
\description{
Calculates a hierarchical scale decomposition of topographic position index
}
\note{
if win  = "circle" units are distance, if win = "rectangle" untis are number of cells
}
\examples{
\dontrun{ 
  library(raster)
  data(elev)
  hsp27 <- hsp(elev, 3, 27, 4, normalize = TRUE)
  plot(hsp27)
 }

}
\references{
Murphy M.A., J.S. Evans, and A.S. Storfer (2010) Quantify Bufo boreas connectivity in Yellowstone National Park with landscape genetics. Ecology 91:252-261
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
