% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{Plot Spatial K nearest neighbor}
\usage{
knn(y, x, k = 1, d = NULL, ids = NULL, indexes = FALSE)
}
\arguments{
\item{y}{Spatial points or polygons object or coordinates matrix}

\item{x}{Spatial points or polygons object or coordinates matrix}

\item{k}{Number of neighbors}

\item{d}{Optional search radius}

\item{ids}{Optional column of ID's in x}

\item{indexes}{(FALSE/TRUE) Return row indexes of x neighbors}
}
\value{
A data.frame with row indexes (optional), rownames, ids (optional) and distance of k
}
\description{
Find K nearest neighbors for two spatial objects
}
\note{
Finds nearest neighbor in x based on y and returns rownames, index and distance,
      If ids is NULL, rownames of x are returned. If coordinate matrix provided, 
      columns need to be ordered [X,Y]. If a radius for d is specified than a maximum 
      search radius is imposed. If no neighbor is found, a neighbor is not returned
}
\examples{
library(sp)
data(meuse)
  coordinates(meuse) <- ~x+y

idx <- sample(1:nrow(meuse), 10) 
  pts <- meuse[idx,]
  meuse <- meuse[-idx,]
    meuse$IDS <- 1:nrow(meuse)

# Find 2 neighbors in meuse
( nn <- knn(pts, meuse, k=2, ids = "IDS", indexes = TRUE) )
   plot(pts, pch=19, main="KNN")
     points(meuse[nn[,1],], pch=19, col="red")

# Using coordinate matrices
y <- coordinates(pts)
x <- coordinates(meuse)
knn(y, x, k=2)
  
}
\seealso{
\code{\link[RANN]{nn2}} for details on search algorithm
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
