% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwCorr.R
\name{mwCorr}
\alias{mwCorr}
\title{Dutilleul moving window bivariate raster 
       correlation}
\usage{
mwCorr(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version 
of the function}
}
\description{
A bivarate raster corrlation using Dutilleul's 
             modified t-test
}
