% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.in.poly.R
\name{point.in.poly}
\alias{point.in.poly}
\title{Point and Polygon Intersect}
\usage{
point.in.poly(...)
}
\arguments{
\item{...}{arguments passed to sf::st_intersection}
}
\value{
NA
}
\description{
Intersects point and polygon feature classes and adds polygon 
             attributes to points
}
\examples{
 \dontrun{
  sf::st_intersection()
}

}
