% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.loess.boot.R
\name{print.loess.boot}
\alias{print.loess.boot}
\title{Print Loess bootstrap model}
\usage{
\method{print}{loess.boot}(x, ...)
}
\arguments{
\item{x}{Object of class loess.boot}

\item{...}{Ignored}
}
\description{
print method for class "loess.boot"
}
