% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.downscale.R
\name{raster.downscale}
\alias{raster.downscale}
\title{Raster Downscale}
\usage{
raster.downscale(
  x,
  y,
  p = NULL,
  n = NULL,
  scatter = FALSE,
  samp.type = c("random", "regular"),
  ...
)
}
\arguments{
\item{x}{A terra SpatRaster object representing independent variable(s)}

\item{y}{A terra SpatRaster object representing dependent variable}

\item{p}{Percent sample size (default NULL)}

\item{n}{Fixed sample size (default NULL)}

\item{scatter}{(FALSE/TRUE) Optional scatter plot}

\item{samp.type}{Type of sample, options are "random" and "regular"}

\item{...}{Additional arguments passed to predict}
}
\value{
A list object containing:
\itemize{ 
\item  downscale downscaled terra SpatRaster object
\item  model     rlm model object 
\item  MSE       Mean Square Error
\item  AIC       Akaike information criterion
}
}
\description{
Downscales a raster to a higher resolution raster using 
             a robust regression
}
\examples{
\dontrun{
  library(geodata)
  library(terra)

  # Download example data (requires geodata package)
    elev <- geodata::elevation_30s(country="SWZ",  path=tempdir())
	  slp <- terrain(elev, v="slope")
    tmax <- geodata::worldclim_country(country="SWZ", var="tmax", 
                                       path=tempdir())
      tmax <- crop(tmax[[1]], ext(elev))
  
  # Downscale temperature
  x=c(elev,slp)
    names(x) <- c("elev","slope")
  y=tmax
    names(y) <- c("tmax")
  
  tmax.ds <- raster.downscale(x, y, scatter=TRUE)
    opar <- par(no.readonly=TRUE)
      par(mfrow=c(2,2))
        plot(tmax, main="Temp max")
        plot(x[[1]], main="elevation")
	    plot(x[[2]], main="slope")
        plot(tmax.ds$downscale, main="Downscaled Temp max")
    par(opar)
   
}
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
