% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.na.omit.R
\name{sp.na.omit}
\alias{sp.na.omit}
\title{sp na.omit}
\usage{
sp.na.omit(...)
}
\arguments{
\item{...}{arguments passed to stats::na.omit}
}
\value{
NA
}
\description{
Removes row or column NA's in sp object
}
