% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cross.cor.R
\name{summary.cross.cor}
\alias{summary.cross.cor}
\title{Summary of spatial cross correlation}
\usage{
\method{summary}{cross.cor}(object, ...)
}
\arguments{
\item{object}{Object of class cross.cor}

\item{...}{Ignored}
}
\description{
summary method for class "cross.cor"
}
