% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.loess.boot.R
\name{summary.loess.boot}
\alias{summary.loess.boot}
\title{Summarizing Loess bootstrap models}
\usage{
\method{summary}{loess.boot}(object, ...)
}
\arguments{
\item{object}{Object of class loess.boot}

\item{...}{Ignored}
}
\value{
same as print lowess.boot data.frame including;
\itemize{
\item nreps        Number of bootstrap replicates
\item confidence   Confidence interval (region)
\item span         alpha (span) parameter used loess fit
\item degree       polynomial degree used in loess fit
\item normalize    Normalized data (TRUE/FALSE)
\item family       Family of statistic used in fit
\item parametric   Parametric approximation (TRUE/FALSE)
\item surface      Surface fit, see loess.control
\item data         data.frame of x,y used in model
\item fit          data.frame including:
\enumerate{
\item x - Equally-spaced x index (see NOTES)
\item y.fit - loess fit
\item up.lim - Upper confidence interval
\item low.lim - Lower confidence interval
\item stddev - Standard deviation of loess fit at each x value
}
}
}
\description{
Summary method for class "loess.boot".
}
