% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.stats.R
\name{zonal.stats}
\alias{zonal.stats}
\title{zonal.stats}
\usage{
zonal.stats(...)
}
\arguments{
\item{...}{arguments passed to terra::extract}
}
\value{
NA
}
\description{
Polygon zonal statistics of a raster
}
\examples{
 \dontrun{
  terra::extract()
}

}
